Per review comments: refactor based on arch
