+++ onnxruntime/core/common/cpuid_info.h
@@ -34,32 +34,76 @@ class CPUIDInfo {
   // ARM 
   bool HasArmNeonDot() const { return has_arm_neon_dot_; }
 
+  uint32_t GetCurrentCoreIdx() const;
+
   /**
    * @return CPU core micro-architecture running the current thread
   */
-  int32_t GetCurrentUarch() const;
+  int32_t GetCurrentUarch() const {
+    if (core_uarchs_.empty()) {
+      return -1;
+    }
+
+    uint32_t coreIdx = GetCurrentCoreIdx();
+    if (coreIdx >= core_uarchs_.size()) {
+      return -1;
+    }
+    return core_uarchs_[coreIdx];
+  }
 
   /**
    * @return CPU core micro-architecture
   */
-  int32_t GetCoreUarch(uint32_t coreId) const;
+  int32_t GetCoreUarch(uint32_t coreId) const {
+    if (coreId >= core_uarchs_.size()) {
+      return -1;
+    }
+    return core_uarchs_[coreId];
+  }
 
   /**
   * @brief Some ARMv8 power efficient core has narrower 64b load/store
   *        that needs specialized optimiztion in kernels
   * @return whether the indicated core has narrower load/store device
   */
-  bool IsCoreArmv8NarrowLd(uint32_t coreId) const;
+  bool IsCoreArmv8NarrowLd(uint32_t coreId) const {
+    if (coreId >= is_armv8_narrow_ld_.size()) {
+      return false;
+    }
+    return is_armv8_narrow_ld_[coreId];
+  }
   
   /**
   * @brief Some ARMv8 power efficient core has narrower 64b load/store
   *        that needs specialized optimiztion in kernels
   * @return whether the current core has narrower load/store device
   */
-  bool IsCurrentCoreArmv8NarrowLd() const;
+  bool IsCurrentCoreArmv8NarrowLd() const {
+    if (is_armv8_narrow_ld_.empty()) {
+      return false;
+    }
+
+    uint32_t coreIdx = GetCurrentCoreIdx();
+    if (coreIdx >= is_armv8_narrow_ld_.size()) {
+      return false;
+    }
+    return is_armv8_narrow_ld_[coreIdx];
+  }
+
 
  private:
-  CPUIDInfo();
+  CPUIDInfo() {
+#ifdef CPUIDINFO_ARCH_X86
+    X86Init();
+#elif defined(CPUIDINFO_ARCH_ARM)
+#ifdef __linux__
+    ArmLinuxInit();
+#elif defined(_WIN32)
+    ArmWindowsInit();
+#endif /* (arm or arm64) and windows */
+#endif
+
+  }
   bool has_avx_{false};
   bool has_avx2_{false};
   bool has_avx512f_{false};
@@ -80,6 +124,18 @@ class CPUIDInfo {
 #if (defined(CPUIDINFO_ARCH_X86) || defined(CPUIDINFO_ARCH_ARM)) && defined(CPUINFO_SUPPORTED)
   bool pytorch_cpuinfo_init_{false};
 #endif
+
+#ifdef CPUIDINFO_ARCH_X86
+  void X86Init();
+#elif defined(CPUIDINFO_ARCH_ARM)
+#ifdef __linux__
+  void ArmLinuxInit();
+#elif defined(_WIN32)
+  void ArmWindowsInit();
+#endif /* (arm or arm64) and windows */
+#endif
+
+
   bool has_arm_neon_dot_{false};
 };
